rename session manager -> storage session manager
