+++ src/py42/_internal/dependency_containers.py
@@ -11,7 +11,7 @@ from py42._internal.login_provider_factories import (
 from py42._internal.modules import archive as archive_module, security as sec_module
 from py42._internal.session import Py42Session
 from py42._internal.session_factory import SessionFactory
-from py42._internal.session_manager import StorageSessionManager
+from py42._internal.storage_session_manager import StorageSessionManager
 
 
 class AuthorityDependencies(object):
