+++ tests/_internal/test_storage_session_manager.py
@@ -2,7 +2,7 @@ import pytest
 
 from py42._internal.login_providers import LoginProvider
 from py42._internal.session_factory import SessionFactory
-from py42._internal.session_manager import StorageSessionManager
+from py42._internal.storage_session_manager import StorageSessionManager
 
 
 def get_session_managers(session_factory):
@@ -20,14 +20,14 @@ def login_provider(mocker):
 
 
 class TestStorageSessionManager(object):
-    def test_get_session_calls_session_factory_with_login_provider(
+    def test_get_storage_session_calls_session_factory_with_login_provider(
         self, session_factory, login_provider
     ):
         storage_session_manager = StorageSessionManager(session_factory)
         storage_session_manager.get_storage_session(login_provider)
         assert session_factory.create_storage_session.call_count == 1
 
-    def test_get_session_with_multiple_calls_calls_factory_only_once(
+    def test_get_storage_session_with_multiple_calls_calls_factory_only_once(
         self, session_factory, login_provider
     ):
         storage_session_manager = StorageSessionManager(session_factory)
@@ -35,7 +35,7 @@ class TestStorageSessionManager(object):
         storage_session_manager.get_storage_session(login_provider)
         assert session_factory.create_storage_session.call_count == 1
 
-    def test_get_session_with_multiple_calls_returns_same_session(
+    def test_get_storage_session_with_multiple_calls_returns_same_session(
         self, session_factory, login_provider
     ):
         storage_session_manager = StorageSessionManager(session_factory)
@@ -43,7 +43,7 @@ class TestStorageSessionManager(object):
         session2 = storage_session_manager.get_storage_session(login_provider)
         assert session1 is session2
 
-    def test_get_session_raises_exception_when_factory_raises_exception(
+    def test_get_storage_session_raises_exception_when_factory_raises_exception(
         self, session_factory, login_provider
     ):
         def mock_get_session(provider, **kwargs):
@@ -58,7 +58,7 @@ class TestStorageSessionManager(object):
         expected_message = "Failed to create or retrieve session, caused by: Mock error!"
         assert e.value.args[0] == expected_message
 
-    def test_get_session_get_saved_session_initially_returns_none(self, session_factory):
+    def test_get_storage_session_get_saved_session_initially_returns_none(self, session_factory):
         storage_session_manager = StorageSessionManager(session_factory)
         assert storage_session_manager.get_saved_session_for_url("TEST-URI") is None
 
@@ -70,7 +70,7 @@ class TestStorageSessionManager(object):
         storage_session_manager.get_storage_session(login_provider)
         assert storage_session_manager.get_saved_session_for_url("TEST-URI") is not None
 
-    def test_get_session_calls_create_session_only_once_for_given_login_provider(
+    def test_get_storage_session_calls_create_session_only_once_for_given_login_provider(
         self, session_factory, login_provider, mocker
     ):
         storage_session_manager = StorageSessionManager(session_factory)
@@ -79,7 +79,7 @@ class TestStorageSessionManager(object):
         storage_session_manager.get_storage_session(login_provider)
         assert storage_session_manager.create_storage_session.call_count == 1
 
-    def test_get_session_calls_get_saved_session_for_url_if_session_already_created(
+    def test_get_storage_session_calls_get_saved_session_for_url_if_session_already_created(
         self, session_factory, login_provider, mocker
     ):
         storage_session_manager = StorageSessionManager(session_factory)
